add crc32

+++ Utility/include/crc32.h
@@ -3,8 +3,7 @@
 
 #include <stdint.h>
 
-uint32_t
-crc32(uint32_t crc, const void *buf, size_t size);
+uint32_t crc32(uint32_t crc, const void *buf, unsigned size);
 
 #endif /* CRC32_H */
